+++ src/index.ts
@@ -7,11 +7,12 @@ import { createCommand } from "./commands/create.js";
 import { completionsCommand } from "./commands/completions.js";
 import { listCommand } from "./commands/list.js";
 import { archiveCommand } from "./commands/archive.js";
+import { hideCommand } from "./commands/hide.js";
 import { editCommand } from "./commands/edit.js";
 import { todayCommand } from "./commands/today.js";
 
 const program = new Command();
-program.name("habitxt").description("Text-based habit tracker").version("1.1.1");
+program.name("habitxt").description("Text-based habit tracker").version("1.1.2");
 
 doCommand(program);
 showCommand(program);
@@ -20,6 +21,7 @@ createCommand(program);
 completionsCommand(program);
 listCommand(program);
 archiveCommand(program);
+hideCommand(program);
 editCommand(program);
 todayCommand(program);